refactor: QuizId 부모로 끌어올리기 적용
