+++ src/pages/RootTemplate/RootTemplate.tsx
@@ -3,7 +3,6 @@ import { NAVBAR_PAGE } from '@constants/path';
 import { Suspense, useMemo } from 'react';
 import { Outlet, useLocation } from 'react-router-dom';
 import styles from './rootTemplate.module.scss';
-import { QuizProvider } from 'src/context/QuizProvider';
 import Spinner from '@components/@common/spinner';
 
 const RootTemplate = () => {
@@ -15,14 +14,12 @@ const RootTemplate = () => {
 	);
 
 	return (
-		<QuizProvider>
-			<div className={styles.container}>
-				<Suspense fallback={<Spinner />}>
-					<Outlet />
-				</Suspense>
-				{navbar}
-			</div>
-		</QuizProvider>
+		<div className={styles.container}>
+			<Suspense fallback={<Spinner />}>
+				<Outlet />
+			</Suspense>
+			{navbar}
+		</div>
 	);
 };
